Update data_writer
This project is moving slowly due to school load.  I fixed a few dumb mistakes on my part surrounding the image handling and now works as it should taking only the PTesseract output and converting to binary to be streamed to the CSV file.  Unfortunately code now produces an HTML error code 429 which means it is receiving too many requests to the URL of the first submission found.  This means either the program is iterating too fast or is incorrectly iterating too many times over the same submission.  Will be fixed when I have the time.